ai_service/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
pydantic
pydantic-settings
httpx
redis
python-multipart
